Revert "feat(dashboard): persistent assistant dock with per-route suggestions (#3386)"
This reverts commit cf8422003271c9f6c1b7919ce0f472129299f3b2.
