+++ client/dashboard/src/components/insights-context.ts
@@ -1,4 +1,3 @@
-import type { InsightsSuggestion } from "@/lib/insights-suggestions";
 import type { ElementsConfig } from "@gram-ai/elements";
 import { createContext, useContext } from "react";
 
@@ -11,7 +10,11 @@ export interface InsightsConfigOptions {
   mcpConfig?: Omit<ElementsConfig, "variant" | "welcome" | "theme">;
   title?: string;
   subtitle?: string;
-  suggestions?: InsightsSuggestion[];
+  suggestions?: Array<{
+    title: string;
+    label: string;
+    prompt: string;
+  }>;
   contextInfo?: string;
   /** Hide the trigger button (e.g., when logs are disabled on this page). */
   hideTrigger?: boolean;
@@ -28,6 +31,11 @@ export interface InsightsContextValue {
    *  Fires once per call — intended for "Explore with AI" CTAs that should
    *  drop the user straight into a running conversation. */
   sendPrompt: (prompt: string) => void;
+  /** Monotonically incrementing counter the provider bumps when the
+   *  trigger should play its one-shot spin animation (e.g. when the
+   *  keyboard shortcut fires). Starts at 0 — consumers should ignore the
+   *  initial value and only react to changes. */
+  triggerSpinKey: number;
 }
 
 export const InsightsContext = createContext<InsightsContextValue>({
@@ -36,6 +44,7 @@ export const InsightsContext = createContext<InsightsContextValue>({
   setIsExpanded: () => {},
   setOverride: () => {},
   sendPrompt: () => {},
+  triggerSpinKey: 0,
 });
 
 /**
