+++ client/dashboard/src/components/observe/LogsTools.tsx
@@ -1,7 +1,7 @@
 import { EditServerNameDialog } from "@/pages/hooks/EditServerNameDialog";
 import { EnableLoggingOverlay } from "@/components/EnableLoggingOverlay";
 import { EnterpriseGate } from "@/components/enterprise-gate";
-import { InsightsConfig } from "@/components/insights-dock";
+import { InsightsConfig } from "@/components/insights-sidebar";
 import { ObservabilitySkeleton } from "@/components/ObservabilitySkeleton";
 import { ErrorAlert } from "@/components/ui/alert";
 import { Button } from "@/components/ui/button";
