+++ client/dashboard/src/components/sidebar-footer-action.tsx
@@ -3,67 +3,28 @@ import type { LucideIcon } from "lucide-react";
 import type { ReactNode } from "react";
 import { Link } from "react-router";
 
-const ACTION_CLASS =
-  "flex min-w-0 flex-1 items-center gap-2 hover:no-underline group-data-[collapsible=icon]:flex-none group-data-[collapsible=icon]:justify-center";
-
 /**
  * A raised, bordered link card that stands out from the sidebar surface — used
  * to stack standout actions (e.g. "Finish setup", "Organization settings") just
  * above the user bar in the sidebar footer. Collapses to an icon when the
- * sidebar does. Pass `to` for a navigation action or `onClick` for an in-place
- * action (e.g. restoring a dismissed surface). `trailing` renders a secondary
- * control (e.g. a restore button) on the card's right edge, outside the main
- * action's click area.
+ * sidebar does. `trailing` renders a secondary control (e.g. a restore button)
+ * on the card's right edge, outside the link's click area.
  */
-type SidebarFooterActionProps = {
-  icon: LucideIcon;
-  label: string;
-  className?: string;
-  contentClassName?: string;
-  trailing?: ReactNode;
-} &
-  // Exactly one of `to` (navigation) or `onClick` (in-place action) — both
-  // optional would let a caller render a card that does nothing.
-  ({ to: string; onClick?: never } | { to?: never; onClick: () => void });
-
 export function SidebarFooterAction({
   to,
-  onClick,
   icon: Icon,
   label,
   className,
   contentClassName,
   trailing,
-}: SidebarFooterActionProps): JSX.Element {
-  const actionContent = (
-    <>
-      <Icon className="size-4 shrink-0" strokeWidth={1.75} />
-      <span className="truncate group-data-[collapsible=icon]:hidden">
-        {label}
-      </span>
-    </>
-  );
-
-  let action: ReactNode;
-  if (to) {
-    action = (
-      <Link to={to} title={label} className={ACTION_CLASS}>
-        {actionContent}
-      </Link>
-    );
-  } else {
-    action = (
-      <button
-        type="button"
-        onClick={onClick}
-        title={label}
-        className={cn(ACTION_CLASS, "text-left")}
-      >
-        {actionContent}
-      </button>
-    );
-  }
-
+}: {
+  to: string;
+  icon: LucideIcon;
+  label: string;
+  className?: string;
+  contentClassName?: string;
+  trailing?: ReactNode;
+}): JSX.Element {
   return (
     <div
       className={cn(
@@ -77,7 +38,16 @@ export function SidebarFooterAction({
           contentClassName,
         )}
       >
-        {action}
+        <Link
+          to={to}
+          title={label}
+          className="flex min-w-0 flex-1 items-center gap-2 hover:no-underline group-data-[collapsible=icon]:flex-none group-data-[collapsible=icon]:justify-center"
+        >
+          <Icon className="size-4 shrink-0" strokeWidth={1.75} />
+          <span className="truncate group-data-[collapsible=icon]:hidden">
+            {label}
+          </span>
+        </Link>
         {trailing}
       </div>
     </div>
