+++ client/dashboard/src/pages/org/OrgAuditLogs.tsx
@@ -2,8 +2,10 @@ import { useQueryState } from "nuqs";
 import type { MCPServerEntry } from "@gram-ai/elements";
 import { recommended } from "@gram-ai/elements/plugins";
 import { RequireScope } from "@/components/require-scope";
-import { InsightsConfig, InsightsProvider } from "@/components/insights-dock";
-import { INSIGHTS_SUGGESTIONS } from "@/lib/insights-suggestions";
+import {
+  InsightsConfig,
+  InsightsProvider,
+} from "@/components/insights-sidebar";
 import { Page } from "@/components/page-layout";
 import { Heading } from "@/components/ui/heading";
 import { Button } from "@/components/ui/button";
@@ -554,7 +556,26 @@ function AuditLogsInsightsWrapper({ children }: { children: React.ReactNode }) {
       mcpConfig={mcpConfig}
       title="Audit Log Insights"
       subtitle="Ask about organization activity, changes, and audit events."
-      suggestions={INSIGHTS_SUGGESTIONS["org/audit-logs"]}
+      suggestions={[
+        {
+          title: "Recent changes",
+          label: "What changed recently?",
+          prompt:
+            "Summarize the most significant recent changes across the organization based on the audit logs.",
+        },
+        {
+          title: "Security review",
+          label: "Security-relevant events",
+          prompt:
+            "What security-relevant events have occurred recently? Look for API key changes, permission modifications, or unusual patterns.",
+        },
+        {
+          title: "Active users",
+          label: "Most active team members",
+          prompt:
+            "Who have been the most active users recently and what kinds of changes have they been making?",
+        },
+      ]}
     >
       {children}
     </InsightsProvider>
